Auto logging on teardown (#293)
* Wait for resource deletions in DeleteResources

This patch makes the method `f.DeleteResources` semantic more
clear by waiting for all resources to be deleted before returning.

This guarantees that once a feature passed to a next phase
resources were actually deleted.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>

* Auto-logging on feature teardown

The existing tests framework in eventing has the nice feature
of logging resources that are tracked during a test and that's
helpful for debugging test issues.

This PR is adding support for the same feature, by automatically
prepending logging steps to the teardown phase of every feature.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>

* Update pkg/environment/logging.go

Co-authored-by: Chris Suszynski <krzysztof.suszynski@gmail.com>

* Update pkg/environment/logging.go

* Add logging steps only when test fails

Co-authored-by: Chris Suszynski <krzysztof.suszynski@gmail.com>