Allow verify service readiness when istio is installed
The WaitForServiceReady was deploying a job without using the job
package which means that the pod created out of the job is not
annotated with istio injection annotation.

In addition, jobs with istio need to quit their sidecar when
they ends so that the job can be considered "successfull".

Unfortunately, moving it to use job created a bigger problem
caused by import cycles, since job was depending from k8s
package and k8s package would have depended on job package.
To solve that problem, I needed to restructure the package
structure so that k8s package dependes on lower level packages
like `job`, `pod`, etc.

I think the k8s package and functions should be all moved to
their respective resource package because it now contains
a lot of unrelated content and it isn't clear what is the
difference between `k8s` and packages under `resources/*`.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>