Refactor image provider code
The code in images.go was using the pattern of defining methods on keys
to get objects from the context, instead of the common pattern in
Knative to have `getX(ctx context.Context) X` and
`withX(ctx context.Context, x X) context.Context`.

In addition, we were using a channel instead of a normal list to collect
packages, but we only need a list there.

Now, we are also publish multiple images in parallel instead of
sequencially.

Improved the check for checking whether the provided path for the
ko image publisher is a Go package (see code comment for details)

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>