Add support for feature-level prerequisites
When testing TLS we are facing the challenge of defining
tests which should be executed based on the feature flag
value.

For example, when we will add tests for [1], we will need
to execute the test only when the transport-encryption
feature flag is set to one of permissive or strict.

[1] https://github.com/knative/eventing/issues/6932

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>