[release-1.9] Remove resources when terminating the environment (#532)
* Remove resources when terminating the environment

When calling env.Finish() the resources that have been created
by such environment should be removed.

So far, we were relying on the namespace deletion to clean
up resources, however, if a precreated namespace is used
for multiple tests and multiple environments, the cluster
is full of resources that are not needed anymore since
those tests were successfully run.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>

* Remove refFailedDeletion since it's useless

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>

* Remove wrong comment

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>

---------

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>
Co-authored-by: Pierangelo Di Pilato <pierdipi@redhat.com>